refactored backend code, and set up user endpoint for scheduler
