+++ api/scheduler/scheduler_functions.js
@@ -1,5 +1,8 @@
 const DSbaseURL = "https://wildfirewatch.herokuapp.com";
 const axios = require("axios");
+const Locations = require("../../models/locations/locations-model");
+const { alertMessage } = require("../../sms/twilio");
+const push = require("../../push/helper");
 
 const getAmericaFires = () => {
   return axios.get(`${DSbaseURL}/fpfire`).then(res => {
@@ -46,7 +49,111 @@ const haversineDistance = (latlngA, latlngB, isMiles) => {
   return finalDistance;
 };
 
+
+async function sendSmsAndPushNotifications(user_id = null) {
+  try {
+    console.log("Running Scheduler");
+
+    let USAfires = await getAmericaFires();
+
+    var locations = [];
+
+    if (user_id === null) {
+      locations = await Locations.findAll();
+    } else {
+      locations = await Locations.findBy(location => location.user_id == user_id);
+    }
+
+    let alertLocations = [];
+
+    locations.forEach(location => {
+
+      var nrFires = 0;
+      var closestFireName = '';
+      var closestDistance = -1;
+
+      if (location.latitude && location.longitude) {
+        USAfires.forEach(fireObj => {
+          const fireLocation = fireObj.location;
+          const fireName = fireObj.name;
+
+          let distance = haversineDistance(
+            [location.latitude, location.longitude],
+            [fireLocation[1], fireLocation[0]],
+            true
+          );
+
+          if (distance <= location.radius) {
+            nrFires += 1;
+
+            if (closestDistance === -1 || distance < closestDistance) {
+              closestFireName = fireName;
+              closestDistance = distance;
+            }
+          }
+        })
+
+        if (nrFires > 0) {
+          alertLocations.push([location, closestFireName, closestDistance, nrFires]);
+        }
+      }
+    });
+
+    // console.log(`Found ${alertLocations.length} location / fire instances`);
+
+    // alertLocations = new Set(alertLocations);
+    console.log(`Going to send ${alertLocations.length} alerts`);
+
+    alertLocations.forEach(async alertLocation => {
+      const alertLoc = alertLocation[0];
+      const closestFireName = alertLocation[1];
+      const closestDistance = Math.round(alertLocation[2]);
+      const nrFires = alertLocation[3];
+
+      var body = '';
+
+      if (nrFires === 1) {
+        body = `There is ${nrFires} fire within ${alertLoc.radius} miles of ${alertLoc.address}. The fire, ${closestFireName}, is ${closestDistance} miles from your location.`;
+      } else {
+        body = `There are ${nrFires} fires within ${alertLoc.radius} miles of ${alertLoc.address}. The closest fire, ${closestFireName}, is ${closestDistance} miles from your location.`;
+      }
+
+      // const body = `There are ${nrFires} fires within ${alertLoc.radius} miles of ${alertLoc.address}. The closest fire, ${closestFireName}, is ${closestDistance} miles from your location.`;
+      console.log(`notification_timer: ${alertLoc.notification_timer}`)
+
+      console.log("User id: " + alertLoc.user_id);
+      console.log("receive_sms: " + alertLoc.receive_sms);
+      console.log("receive_push: " + alertLoc.receive_push);
+
+      // if (alertLoc.notification_timer === 0) {
+      if (alertLoc.receive_sms) {
+        alertMessage(alertLoc.cell_number, body);
+      }
+
+      if (alertLoc.receive_push) {
+        push(alertLoc.user_id, {
+          title: "Wildfire Notification",
+          body: body
+        });
+      }
+      // }
+
+      if (alertLoc.notification_timer === 12) {
+        await Locations.update(alertLoc.id, { notification_timer: 0 });
+      } else {
+        await Locations.update(alertLoc.id, {
+          notification_timer: alertLoc.notification_timer + 1
+        });
+      }
+    });
+  } catch (error) {
+    console.log(error);
+  }
+}
+
+
 module.exports = {
   getAmericaFires,
-  haversineDistance
+  haversineDistance,
+  sendSmsAndPushNotifications
 };
