+++ models/locations/locations-model.js
@@ -32,7 +32,21 @@ function findAll() {
 }
 
 function findBy(filter) {
-  return db("locations").where(filter);
+  return db("locations")
+    .where(filter)
+    .join("users", "users.id", "locations.user_id")
+    .select(
+      "locations.id",
+      "locations.user_id",
+      "locations.address",
+      "locations.latitude",
+      "locations.longitude",
+      "locations.radius",
+      "locations.notification_timer",
+      "users.cell_number",
+      "users.receive_sms",
+      "users.receive_push"
+    );
 }
 
 async function add(location) {
